pyproject.toml
[tool.poetry]
name = "youtube-video-resumer"
version = "0.1.0"
description = ""
authors = ["Rafael <rafaeloc.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
pytubefix = "^8.3.0"
openai = "^1.53.0"
ffmpeg-python = "^0.2.0"
llama-index = "^0.11.22"
llama-hub = "^0.0.79.post1"
assemblyai = "^0.34.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
resume = "youtube_video_resumer.main:resume"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
